% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_binary}
\alias{build_binary}
\title{Build binary packages and place them in the repository}
\usage{
build_binary(packages, repo = default_repo())
}
\arguments{
\item{packages}{Names of the packages.}

\item{repo}{Path to the repository. Defaults to [default_repo()].}
}
\description{
Build binary packages and place them in the repository
}
