% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-api.R
\name{cache_summary}
\alias{cache_summary}
\alias{cache_list}
\alias{cache_find}
\alias{cache_get_file}
\alias{cache_delete_files}
\alias{cache_add_file}
\title{Functions to query and manipulate the package cache}
\usage{
cache_summary(cachepath = NULL)

cache_list(cachepath = NULL)

cache_find(cachepath = NULL, ...)

cache_get_file(cachepath = NULL, target, ...)

cache_delete_files(cachepath = NULL, ...)

cache_add_file(cachepath = NULL, file, relpath = dirname(file), ...)
}
\arguments{
\item{cachepath}{Path of the cache. By default the cache directory is in
\code{R-pkg}, within the user's cache directory.
See \code{\link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir()}}.}

\item{...}{Extra named arguments to select the package file.}

\item{target}{Path where the selected file is copied.}

\item{path}{The relative path of the file within the cache.}
}
\description{
\code{cache_summary()} returns a short summary of the state of the cache,
e.g. the number of files and their total size. It returns a named list.

\code{cache_list()} lists all files in the cache. It returns a \code{tibble}.

\code{cache_find()} finds all files in the cache that match the specified
attributes. It returns a \code{tibble}.

\code{cache_get_file()} copied a file out of the cache into the specified
path. If no file is found, then it returns \code{NULL}. Otherwise it returns
(invisibly) the tibble of all selected files. If multiple
files match the specified attributes, then the first one is copied to
the \code{target} path.

\code{cache_delete_files()} deletes the selected files from the caches.
Without any arguments, all files are deleted.

\code{cache_add_file} adds a file to the cache.
}
