% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-cache.R
\name{cmc__async_ensure_cache}
\alias{cmc__async_ensure_cache}
\title{Load the cache, asynchronously, with as little work as possible}
\usage{
cmc__async_ensure_cache(self, private, max_age)
}
\arguments{
\item{self}{self}

\item{private}{private self}

\item{max_age}{Maximum age allowed to consider the data current.}
}
\value{
Metadata.
}
\description{
\enumerate{
\item If it is already loaded, and fresh return it.
\item Otherwise try the replica RDS.
\item Otherwise try the primary RDS.
\item Otherwise try the primary PACKAGES files.
\item Otherwise update the replica PACKAGES files,
the replica RDS, and then also the primary PACKAGES and RDS.
}
}
\keyword{internal}
