% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-cache.R
\name{cmc__load_primary_pkgs}
\alias{cmc__load_primary_pkgs}
\title{Load metadata from the primary cache's PACKAGES files}
\usage{
cmc__load_primary_pkgs(self, private, max_age)
}
\arguments{
\item{self}{self}

\item{private}{private self}

\item{max_age}{Max age to consider the files current.}
}
\value{
Metadata.
}
\description{
If they are not available, or outdated, it throws an error.
Otherwise they are copied to the replica cache, and then used
to create the RDS file. The RDS file is then written back to the
primary cache and also loaded.
}
\keyword{internal}
