% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-cache.R
\name{cmc__load_primary_rds}
\alias{cmc__load_primary_rds}
\title{Load the metadata from the primary cache's RDS file}
\usage{
cmc__load_primary_rds(self, private, max_age)
}
\arguments{
\item{self}{Self.}

\item{private}{Private self.}

\item{max_age}{Maximum age allowed for the RDS file to be considered
as current.}
}
\value{
Metadata.
}
\description{
If it exists and current, then the replica RDS is updated to it as well,
and the data is returned. Otherwise throws an error.
}
\keyword{internal}
