% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-cache.R
\name{cmc__load_replica_rds}
\alias{cmc__load_replica_rds}
\title{Try to load the package metadata asynchronously, from the replica RDS}
\usage{
cmc__load_replica_rds(self, private, max_age)
}
\arguments{
\item{self}{Self.}

\item{private}{Private self.}

\item{max_age}{Maximum age allowed for the RDS file to be considered
as current.}
}
\value{
The metadata.
}
\description{
If the replica has the RDS data, it is loaded and returned.
Otherwise throws an error.
}
\keyword{internal}
