% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-cache.R
\name{cmc__update_replica_pkgs}
\alias{cmc__update_replica_pkgs}
\title{Update the PACKAGES files in the replica cache}
\usage{
cmc__update_replica_pkgs(self, private)
}
\arguments{
\item{self}{self}

\item{private}{private self}
}
\description{
I.e. download them, if they have changed.
}
\keyword{internal}
