% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-cache.R
\name{cmc__update_replica_rds}
\alias{cmc__update_replica_rds}
\title{Update the replica RDS from the PACKAGES files}
\usage{
cmc__update_replica_rds(self, private)
}
\arguments{
\item{self}{self}

\item{private}{private self}
}
\description{
Also loads it afterwards.
}
\keyword{internal}
