% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-cache.R
\name{cran_deps}
\alias{cran_deps}
\alias{cran_revdeps}
\alias{cran_update}
\alias{cran_list}
\title{Query CRAN(like) package data}
\usage{
cran_deps(packages, dependencies = NA, recursive = TRUE)

cran_revdeps(packages, dependencies = NA, recursive = TRUE)

cran_update()

cran_list(packages = NULL)
}
\arguments{
\item{packages}{Packages to query.}

\item{dependencies}{Dependency types to query. See the \code{dependencies}
parameter of \code{\link[utils:install.packages]{utils::install.packages()}}.}

\item{recursive}{Whether to query recursive dependencies.}
}
\value{
A data frame (tibble) of the dependencies. For \code{cran_deps()}
it includes the queried \code{packages} as well.
}
\description{
It uses CRAN and BioConductor packages.
}
\details{
\code{cran_list()} lists all packages.

\code{cran_update()} updates all metadata. Note that metadata is automatically
updated if it is older than seven days.

\code{cran_deps()} queries packages dependencies.

\code{cran_revdeps()} queries reverse package dependencies.
}
\section{Examples}{
\preformatted{cran_deps("dplyr")
cran_list(c("MASS", dplyr"))
cran_update()
}
}

