% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-resolution.R
\name{get_all_package_dirs}
\alias{get_all_package_dirs}
\title{Work out all package directories}
\usage{
get_all_package_dirs(platforms, rversions)
}
\arguments{
\item{platforms}{Can be a subset of `source`, `windows`, `macos`. The
platforms to download.}

\item{rversions}{The R versions to download.}
}
\description{
Based on the platforms and R versions that we support.
}
\keyword{internal}
