% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-resolution.R
\name{get_cran_extension}
\alias{get_cran_extension}
\title{Package file extension for a platform}
\usage{
get_cran_extension(platform)
}
\arguments{
\item{platform}{Platform.}
}
\value{
File extension.
}
\description{
Package file extension for a platform
}
\keyword{internal}
