% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-releases.R
\name{ghr$add_asset}
\alias{ghr$add_asset}
\title{Add asset got GitHub release}
\usage{
ghr%add_asset(repo, file, tag, name = basename(file))
}
\arguments{
\item{repo}{Repository slug, e.g. \code{cran/cli}.}

\item{file}{Path to file to upload as an asset.}

\item{tag}{Tag name to add the asset to. It must exist on GitHub.}

\item{name}{File name of the asset in the release.}
}
\value{
Response from GitHub as a named list. See
\url{https://docs.github.com/en/rest/releases/assets#upload-a-release-asset}
for the structure.
}
\description{
\code{ghr$async_add_asset()} is the async version of \code{ghr$add_asset()}.
}
\keyword{internal}
