% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-releases.R
\name{ghr$create}
\alias{ghr$create}
\title{Create a GitHub release}
\usage{
ghr$create(
  repo,
  tag,
  description = "",
  draft = FALSE,
  prerelease = FALSE,
  generage_release_notes = FALSE
)
}
\arguments{
\item{repo}{Repository slug, e.g. \code{cran/cli}.}

\item{tag}{Tag name to create a release for. It must exist on GitHub.}

\item{description}{Release description.}

\item{draft}{Whether to create a draft release.}

\item{prerelease}{Whether to create a prerelease.}

\item{generate_release_notes}{Whether to auto-generate release notes.}
}
\value{
Response from GitHub as a named list. See
\url{https://docs.github.com/en/rest/releases/releases#create-a-release}
for the structure.
}
\description{
\code{ghr$async_create()} is an async version of \code{ghr$create()}.
}
\keyword{internal}
