% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-releases.R
\name{ghr$get}
\alias{ghr$get}
\title{Get information about a release, including release assets}
\arguments{
\item{repo}{Repository slug, e.g. \code{"cran/cli"}.}

\item{tag}{Tag to get.}
}
\value{
Named list, see
\url{https://docs.github.com/en/rest/releases/releases#get-a-release}
for the entries.
}
\description{
\code{ghr$async_get} is the async version of \code{ghr$get}.
}
\keyword{internal}
