% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-releases.R
\name{ghr$list}
\alias{ghr$list}
\title{List GitHub releases}
\arguments{
\item{repo}{Repository slug, e.g. \code{"cran/cli"}.}
}
\value{
Data frame with columns:
\itemize{
\item \code{id}: release id,
\item \code{name}: release name, usuaally the version number, possibly with
a \code{v} prefix: \verb{3.6.1} or \code{v3.6.1}, but can be different.
\item \code{tag_name}: usually the same as \code{name}.
\item \code{created_at}: \code{POSIXct} vector.
}
}
\description{
Lists the last 100 releases for a GitHub repository.
}
\details{
\code{ghr$async_list()} is the async version of \code{ghr$list()}.
}
\keyword{internal}
