% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-releases.R
\name{ghr$list_assets}
\alias{ghr$list_assets}
\title{List assets of a release}
\usage{
ghr$list_assets(repo, tag)
}
\arguments{
\item{repo}{Repository slug, e.g. \code{"cran/cli"}.}

\item{tag}{Tag to query.}
}
\value{
Data frame with columns:
\itemize{
\item \code{id}: asset id,
\item \code{name}: file name of the asset,
\item \code{download_url}: download URL,
\item \code{size}: size in bytes,
\item \code{created_at}: \code{POSIXct} vector,
\item \code{updated_at}: \code{POSXct} vector,
\item \code{content_type}: content type of asset.
}
}
\description{
\code{ghr$async_list_assets()} is the async version of \code{ghr$list_assets()}
}
\keyword{internal}
