% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-repo.R
\name{ghrepo}
\alias{ghrepo}
\title{Update a CRAN-like repository of binary packages at GitHub}
\description{
These functions are currently experimental.
}
\details{
\subsection{Update a CRAN-like repository of binary packages at GitHub}{

\code{ghrepo$update()} updates a binary package mirror.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{ghrepo$update(
  repo,
  subdir,
  release_org = "cran",
  source_repo = "https://cran.r-project.org",
  packages = NULL
)
}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\itemize{
\item \code{repo}: GitHub slug, e.g. \code{r-hub/repos}.
\item \code{subdir}: subdirectory in the GitHub repository, where the R package
metadata should be updated. It must exist in the repository.
If it does not have \verb{PACKAGES*} files, then they will be created.
\item \code{release_org}: GitHub organization or user name where the packages
will be published as releases.
\item \code{source_repo}: A CRAN-like repository, where source packages are
taken from.
\item \code{packages}: A character vector of package names to add to the binary
repository, in addition to updating the ones that are already there.
}
}

}
}
\keyword{internal}
