% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-protocol.R
\name{git-protocol}
\alias{git-protocol}
\title{git protocol notes, for developers}
\description{
Assumptions, they might be relaxed or checked for later:
\itemize{
\item The server must speak the smart protocol, version 2.
\item We use HTTP transport, not SSH.
\item The server should have the \code{shallow} capability.
\item The server should have the \code{filter} capability.
}
}
\details{
Improvements needed:
\itemize{
\item Tests.
\item Use async HTTP.
\item Support packfiles with deltas.
\item Optionally send authorization.
\item Better error messages.
\item Better errors for non-existing user, repository, ref, PR, etc.
}
\subsection{Docs and other helpful links:}{
\itemize{
\item \url{https://github.com/git/git/blob/master/Documentation/gitprotocol-common.txt}
\item \url{https://github.com/git/git/blob/master/Documentation/gitprotocol-pack.txt}
\item \url{https://github.com/git/git/blob/master/Documentation/gitprotocol-v2.txt}
\item \url{https://github.com/calebsander/git-internals/blob/part2/src/main.rs}
}
}
}
\keyword{internal}
