% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-protocol.R
\name{git_list_refs}
\alias{git_list_refs}
\title{List references in a remote git repository}
\usage{
git_list_refs(url, prefixes = NULL)
}
\arguments{
\item{url}{Repository URL, e.g. \verb{https://github.com/r-lib/pak.git}.}

\item{prefixes}{If not \code{NULL}, then only references of which one of
\code{prefixes} is a prefix, are listed.}
}
\value{
A list with entries: \code{refs} and \code{caps}.
\code{caps} is a character vector of capabilities advertised by the server.
\code{refs} is a data frame of git refs, it has columns \code{ref} and \code{hash}.
}
\description{
List references in a remote git repository
}
\details{
\itemize{
\item Branches have references named \verb{refs/heads/<branch>}, e.g.
\code{refs/heads/main}.
\item Tags have references named \verb{refs/tags/<tag>}, e.g. \code{refs/tags/v1.0.2}.
\item On GitHub pull requests have references named
\verb{refs/pull/<pr-number>/head}, e.g. \code{refs/pull/37/head}.
For open pull requests there should be a \verb{refs/pull/<pr-number>/merge}
reference as well, which is the branch after the pull request has been
merged.
\item There is a special reference called \code{HEAD}. This points to the default
branch on GitHub.
}
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{git_list_refs("https://github.com/r-lib/filelock.git")
#> $refs
#>                                       ref
#> 1                                    HEAD
#> 2                   refs/heads/cran-1-0-2
#> 3  refs/heads/fix/testthat-error-parallel
#> 4                     refs/heads/gh-pages
#> 5                         refs/heads/main
#> 6                  refs/heads/refcounting
#> 7                       refs/pull/14/head
#> 8                       refs/pull/37/head
#> 9                      refs/pull/37/merge
#> 10                      refs/pull/38/head
#> 11                      refs/pull/40/head
#> 12                       refs/tags/v1.0.2
#>                                        hash
#> 1  0331c49c0b17845468b5f3e2da75680475bc9989
#> 2  dfc615b14dfc030df3f44bd377e752728f76df3d
#> 3  6f4b4bfcc1d3442d63496c7640984c0d9570a05a
#> 4  2956fbdcbab65b75e3347048a9dc18b15fe377d2
#> 5  0331c49c0b17845468b5f3e2da75680475bc9989
#> 6  48e5bbe0e031897c7f0e194edbf6897f813fbaee
#> 7  48e5bbe0e031897c7f0e194edbf6897f813fbaee
#> 8  6f4b4bfcc1d3442d63496c7640984c0d9570a05a
#> 9  74b253543f6d2e7213a8174ebbe8e397b9cc866b
#> 10 c7e58188a55844e6e75cd38cf81388c3452b1259
#> 11 10421f0b7117244002d377d7a9699c2742abdfcc
#> 12 9fdba75a62facaa3e818902f58891166e45eabe9
#> 
#> $caps
#>  [1] "multi_ack"                      "thin-pack"                     
#>  [3] "side-band"                      "side-band-64k"                 
#>  [5] "ofs-delta"                      "shallow"                       
#>  [7] "deepen-since"                   "deepen-not"                    
#>  [9] "deepen-relative"                "no-progress"                   
#> [11] "include-tag"                    "multi_ack_detailed"            
#> [13] "allow-tip-sha1-in-want"         "allow-reachable-sha1-in-want"  
#> [15] "no-done"                        "symref=HEAD:refs/heads/main"   
#> [17] "filter"                         "object-format=sha1"            
#> [19] "agent=git/github-g8119c27a9d8d"
}\if{html}{\out{</div>}}

List only references having a certain prefix:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{git_list_refs("https://github.com/r-lib/filelock.git", "refs/heads/main")
#> $refs
#>               ref                                     hash
#> 1 refs/heads/main 0331c49c0b17845468b5f3e2da75680475bc9989
#> 
#> $caps
#> [1] "agent=git/github-g8119c27a9d8d"     "ls-refs=unborn"                    
#> [3] "fetch=shallow wait-for-done filter" "server-option"                     
#> [5] "object-format=sha1"
}\if{html}{\out{</div>}}
}

\keyword{internal}
