% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-remotes.R
\name{github_username_rx}
\alias{github_username_rx}
\title{Match a GH username}
\usage{
github_username_rx()
}
\description{
\itemize{
\item may only contain alphanumeric characters or hyphens
\item cannot have multiple consecutive hyphens
\item cannot begin or end with a hyphen
\item maximum 39 characters
}
}
\details{
Based on https://github.com/shinnn/github-username-regex
}
\keyword{internal}
