% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-plan.R
\name{install_package_plan}
\alias{install_package_plan}
\title{Perform a package installation plan, as created by pkgdepends}
\usage{
install_package_plan(plan, lib = .libPaths()[[1]], num_workers = 1)
}
\arguments{
\item{plan}{Package plan object, returned by pkgdepends}

\item{lib}{Library directory to install to.}

\item{num_workers}{Number of worker processes to use.}
}
\value{
Information about the installation process.
}
\description{
Perform a package installation plan, as created by pkgdepends
}
