% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-name.R
\name{is_valid_package_name}
\alias{is_valid_package_name}
\title{Check whether a package name is valid}
\usage{
is_valid_package_name(nm)
}
\arguments{
\item{nm}{Potential package name, string of length 1.}
}
\value{
Logical flag. If \code{FALSE}, then the \code{reason} attribute
contains a character string, the explanation why the package name
is invalid. See examples below.
}
\description{
Check whether a package name is valid
}
\examples{
is_valid_package_name("pak")
is_valid_package_name("pkg")
is_valid_package_name("pak\u00e1ge")
is_valid_package_name("good-package")
is_valid_package_name("x")
is_valid_package_name("1stpackage")
is_valid_package_name("dots.")
}
