% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-installed.R
\name{lib_status}
\alias{lib_status}
\title{Status of packages in a ibrary}
\usage{
lib_status(library = .libPaths()[1], packages = NULL)
}
\arguments{
\item{lib}{Path to library.}
}
\value{
Data frame (tibble) the contains data about the packages
installed in the library.
}
\description{
Status of packages in a ibrary
}
