% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep-utils.R
\name{local_deps}
\alias{local_deps}
\title{Dependency refs of a package file or tree}
\usage{
local_deps(path = ".")
}
\arguments{
\item{path}{Path to the package file or tree.}
}
\value{
Tibble with columns:
\itemize{
\item \code{ref} Remote reference.
\item \code{type} Dependency type.
\item \code{package} Package name.
\item \code{op} Version requirement operator. Empty string if there is no
version requirement.
\item \code{version} Version requirement version.
}
}
\description{
Dependency refs of a package file or tree
}
