% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-tar.R
\name{make_untar_process}
\alias{make_untar_process}
\title{Create a tar background process}
\usage{
make_untar_process(
  tarfile,
  files = NULL,
  exdir = ".",
  restore_times = TRUE,
  post_process = NULL
)
}
\arguments{
\item{tarfile}{Tar file.}

\item{files}{Files or regular expressions to set what to extract. if
\code{NULL} then everything is extracted.}

\item{exdir}{Where to extract the archive. It must exist.}

\item{restore_times}{Whether to restore file modification times.}

\item{post_process}{Function to call after the extraction.}
}
\value{
The \link[callr:process]{callr::process} object.
}
\description{
Use an external tar program, if there is a working one, otherwise use
the internal implementation.
}
\details{
When using the internal implementation, we need to start another R
process.
}
\keyword{internal}
