% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast-resolve.R
\name{merge_bioc_data}
\alias{merge_bioc_data}
\title{Merge the BioC package metadata from the various BioC repos}
\usage{
merge_bioc_data(data, repos)
}
\arguments{
\item{data}{Named list of repo metadata. Each entry has two tibbles in
a list: \code{pkgs} and \code{deps}.}

\item{repos}{List with entries: \code{repos}, the actual repository URLs, in
a named character vector, and \code{version}, the R version that belongs
to these BioC repos.}
}
\value{
A list with two entries: \code{pkgs} and \code{deps}, the merged package
tibble and dependency tibble.
}
\description{
This will simplify dependency lookup, and we can use the same algorithm
as for CRAN metadata (or any CRAN-like repo, maybe?).
}
\details{
We add an \code{X-RPKG-BioCRepo} column to the \code{pkgs} data frames, merge them,
and update the \code{idx} columns in the \code{deps} data frames, before merging
them as well.
}
\keyword{internal}
