% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-downloads.R
\name{new_pkg_download_proposal}
\alias{new_pkg_download_proposal}
\title{Create a package download proposal}
\usage{
new_pkg_download_proposal(refs)
}
\arguments{
\item{refs}{Package names or references. See \link{pkg_refs} for the
syntax.}
}
\description{
A package download proposal can be used to query package dependencies
and/or download package files.
}
\details{
Typical workflow to download a set of packages:\preformatted{pkgdl <- new_pkg_download_proposal(refs)
pkgdl$resolve()
pkgdl$download()
pkgdl$get_downloads()
}
}
\examples{
\donttest{
dl <- new_pkg_download_proposal("r-lib/pak")
dl
dl$resolve()
dl$get_resolution()[]
dl$download()
dl$get_downloads()[]
}
}
\seealso{
\link{pkg_download_proposal} for the methods returned R6 object.
}
