% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-installation.R
\name{new_pkg_installation_proposal}
\alias{new_pkg_installation_proposal}
\title{Create a package installation proposal}
\usage{
new_pkg_installation_proposal(
  refs,
  library,
  upgrade = FALSE,
  dependencies = NA
)
}
\arguments{
\item{refs}{Package names or references. See \link{pkg_refs} for the
syntax.}

\item{library}{Path to package library to install to. If it does not
exist, it will be created.}

\item{upgrade}{Whether to update packages (including dependencies)
to their latest available version.}

\item{dependencies}{Whether dependent packages should be installed,
and which ones. Defaults to installing the required (recursive)
dependencies.}
}
\description{
A package installation proposal can be used to install or upgrade
a set of packages in a package library.
}
\details{
Typical workflow to install a set of packages:\preformatted{inst <- new_pkg_installation_proposal(refs, library)
inst$solve()
inst$download()
inst$install()
}
}
\examples{
\donttest{
inst <- new_pkg_installation_proposal("r-lib/cli", library = tempfile())
inst
inst$solve()
inst$draw()
inst$download()
inst$get_install_plan()
inst$install()
}
}
\seealso{
\link{pkg_installation_proposal} for the methods of the returned R6
object.
}
