% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-cache.R
\docType{data}
\name{package_cache}
\alias{package_cache}
\title{A simple package cache}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
package_cache
}
\description{
Fields:
\itemize{
\item \code{fullpath} Full package path.
\item \code{path} Package path, within the repository.
\item \code{package} Package name.
\item \code{url} URL it was downloaded from.
\item \code{etag} ETag for the last dowload, from the given URL.
\item \code{md5} MD5 of the file, to make sure if it has not changed.
}
}
\keyword{internal}
