% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-resolution.R
\name{parse_deps}
\alias{parse_deps}
\title{Parse dependency specifications, vectorized}
\usage{
parse_deps(deps)
}
\arguments{
\item{deps}{Vector of dependency specifications.}
}
\value{
Character vector of dependent packages. Base packages are
  omitted.
}
\description{
Parse dependency specifications, vectorized
}
\keyword{internal}
