% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{pkg_build}
\alias{pkg_build}
\title{Create a binary package from an installed package}
\usage{
pkg_build(
  pkg,
  library = .libPaths()[1],
  flavor = Sys.getenv("PKG_BUILD_FLAVOR"),
  build_number = 1L
)
}
\arguments{
\item{pkg}{Package name.}

\item{library}{Library path.}

\item{flavor}{Platform flavor. Defaults to the \code{PKG_BUILD_FLAVOR}
environment variable. If not \code{NULL} or an empty string, then it is
appended to the platform string with a dash.}

\item{build_number}{An integer number that is added to the file name,
after the version number, to be able to have multiple builds for the
same package version.}
}
\value{
Path to the built package.
}
\description{
The built package will be in the current working directory.
}
\details{
This function is currently experimental.
}
\keyword{internal}
