% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-progress-bar.R
\name{pkgdepends-download-progress-bar}
\alias{pkgdepends-download-progress-bar}
\title{The download progress bar}
\description{
This is a short summary of the design of the download progress bar.

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\details{
\subsection{Package sizes}{

One difficulty here is that we don't always know the sizes of the files
we are downloading. We basically have three cases for downloads:
\itemize{
\item For some, we know the sizes from the extra CRAN metadata. This is
usually correct, unless CRAN has changed the file on the web server.
If we get a size from the web server, then we'll use that.
\item For some, we do not know the size beforehand, but the web server
sends us the correct size, together with the first chunk of data.
This is typically for CRAN packages that we don't have metadata
about, Bioconductor packages, or GitHub packages for which GitHub
already knows the size.
\item For some, we do not know the size, and the web server does not tell
us the size, either. This is typically for GH packages that GH is
creating on the fly.
}

So ideally we would have a different progress bar when the total
size if known.
}

\subsection{Possible information to show:}{
\itemize{
\item Visual progress bar
\item Number of packages: done / in-progress / total
\item Number of bytes: done / total (if known)
\item ETA
\item Download rate, i.e. x kB/s
\item Elapsed time
\item Percent of packages done
\item Percent of bytes done
\item Spinner to show that we are still working, although updating ETA is
good as well...
\item Events, e.g. got a package, downloading, or starting...
}
}

\subsection{All package sizes are known}{\preformatted{230 kB/s (#####      ) 45\% | 14/56 pkgs | ETA ~34s | Getting dplyr
}
\itemize{
\item Rate is informative, serves as an "alive" indicator as well.
\item Bar is visual.
\item Percent is the best quantitative measure.
\item Done/total packages is informative, not so much for the download
but the whole installation.
\item ETA is informative.
\item Event is nice, and it fills the space...
\item We don't show current/total bytes, shoulwd we? Quite wide...
}
}

\subsection{Some package sizes are unknown}{\preformatted{230 kB/s (#####      ) 45\% | 14/56 pkgs | Getting dplyr
}
\itemize{
\item Percent is for the packages, not bytes.
\item We can't really show ETA.
}
}

\subsection{Events}{

This is quite simple currently, we just print the packages we are
"getting" currently, i.e. the packages we received data for in the
last time slot.
}

\subsection{Rate (download "speed")}{

To calculate the rate we need to record when the various data chunks
have arrived. We only do this with a resolution of 1s. We create an
environment called chunks, and for every second since the downloads have
started we record the number of bytes that arrived in that second.

When calculating the rate, we take the data for the last 4 seconds,
and normalize it to per second, while taking into account that the
data for the last (current) second is incomplete.

In the first 3 seconds we use all the available data.
}
}
\keyword{internal}
