% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdepends.R
\docType{package}
\name{pkgdepends-package}
\alias{pkgdepends}
\alias{pkgdepends-package}
\title{pkgdepends: Package Dependency Resolution and Downloads}
\description{
\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css"><script src="../doc/assets/rd.js"></script>}}

pkgdepends is a toolkit for package dependencies, downloads and
installations, to be used in other packages. If you are looking for a
package manager, see \href{https://github.com/r-lib/pak}{pak}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/pkgdepends#readme}
  \item Report bugs at \url{https://github.com/r-lib/pkgdepends/issues}
}

}
\section{Features}{
\itemize{
\item Look up package dependencies recursively.
\item Visualize package dependencies.
\item Download packages and their dependencies.
\item Install downloaded packages.
\item Includes a dependency solver to find a consistent set of
dependencies.
\item Supports CRAN and Bioconductor packages automatically.
\item Supports packages on GitHub.
\item Supports local package file and trees.
\item Supports the \code{Remotes} entry in the \code{DESCRIPTION} file.
\item Caches metadata and downloaded packages via
\href{https://github.com/r-lib/pkgcache}{pkgcache}
\item Performs all downloads and HTTP queries concurrently.
\item Builds and installs packages in parallel.
}
}
\section{Install}{\preformatted{install.packages("pkgdepends")
}
}
\section{Usage}{\preformatted{library(pkgdepends)
}
\subsection{Package references}{

A package reference (ref) specifies a location from which an R package
can be obtained from. Examples:\preformatted{devtools
cran::devtools
bioc::Biobase
r-lib/pkgdepends
https://github.com/r-lib/pkgdepends
local::~/works/shiny
}

See \link[=pkg_refs]{‘Package references’} for details.
}

\subsection{Package dependencies}{

TODO
}

\subsection{Package downloads}{

TODO
}

\subsection{Package installation}{

TODO
}

\subsection{Dependency resolution}{

TODO
}

\subsection{The dependency solver}{

TODO
}

\subsection{Installation plans}{

TODO
}

\subsection{Configuration}{

TODO
}

}
\section{Related}{
\itemize{
\item \href{https://github.com/r-lib/pak}{pak} – R package manager
\item \href{https://github.com/r-lib/pkgcache}{pkgcache} – Metadata and pacakge
cache
\item \href{https://github.com/r-lib/devtools}{devtools} – Tools for R package
developers
}
}
\section{License}{

MIT © RStudio
}
