% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdepends.R
\docType{package}
\name{pkgdepends}
\alias{pkgdepends}
\title{pkgdepends}
\description{
The pkgdepends package
}
\details{
pkgdepends is a toolkit for package dependencies, downloads and
installations, to be used in other packages. If you are looking for a
package manager, see \href{https://github.com/r-lib/pak}{pak}.
}
\section{Features}{
\itemize{
\item Look up package dependencies recursively.
\item Visualize package dependencies.
\item Download packages and their dependencies.
\item Install downloaded packages.
\item Includes a dependency solver to find a consistent set of
dependencies.
\item Supports CRAN and Bioconductor packages automatically.
\item Supports packages on GitHub.
\item Supports local package file and trees.
\item Supports the \code{Remotes} entry in the \code{DESCRIPTION} file.
\item Caches metadata and downloaded packages via
\href{https://github.com/r-lib/pkgcache}{pkgcache}
\item Performs all downloads and HTTP queries concurrently.
\item Builds and installs packages in parallel.
}
}
\section{Install}{\if{html}{\out{<div class="r">}}\preformatted{install.packages("pkgdepends")
}\if{html}{\out{</div>}}
}
\section{Usage}{\if{html}{\out{<div class="r">}}\preformatted{library(pkgdepends)
}\if{html}{\out{</div>}}
\subsection{Package references}{

A package reference (ref) specifies a location from which an R package
can be obtained from. Examples:

For CRAN and Bioconductor packages, just use the package names, with an
optional \verb{cran::} or \verb{bioc::} prefix. GitHub packages are in the
\code{user/repo} format, or an URL to the repo. Local package files or trees
use the \verb{local::<path>} format.

Examples:\preformatted{devtools
cran::devtools
bioc::Biobase
r-lib/pkgdepends
https://github.com/r-lib/pkgdepends
local::~/works/shiny
}

Use \code{\link[=parse_pkg_refs]{parse_pkg_refs()}} to parse package
references.
}

\subsection{Package dependencies}{

To query package dependencies, use
}

\subsection{Package downloads}{

TODO
}

\subsection{Package installation}{

TODO
}

}
\section{Related}{
\itemize{
\item \href{https://github.com/r-lib/pak}{pak} – R package manager
\item \href{https://github.com/r-lib/pkgcache}{pkgcache} – Metadata and pacakge
cache
\item \href{https://github.com/r-lib/devtools}{devtools} – Tools for R package
developers
}
}
\section{License}{

MIT © RStudio
}
