% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-progress-bar.R
\name{pkgplan__show_progress_bar}
\alias{pkgplan__show_progress_bar}
\title{Show / update the progress bar}
\usage{
pkgplan__show_progress_bar(bar)
}
\arguments{
\item{bar}{The progress bar object.}
}
\description{
Show / update the progress bar

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\keyword{internal}
