% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-download.R
\name{remotes__download_github}
\alias{remotes__download_github}
\title{Download and build a package from GitHub}
\usage{
remotes__download_github(self, private, resolution)
}
\arguments{
\item{self}{Self.}

\item{private}{Private self.}

\item{resolution}{Dependency resolution result.}
}
\description{
It supports caching, both for the downloaded zip file, and the
built source package. The algorithm is this:
\itemize{
\item Check if we have a built package in the cache. If yes, do nothing.
\item Check if we have a downloaded zip file in the cache. If yes, build it
into the cached source package.
\item Otherwise download the zip file and build it.
}
}
\details{
The name of the zip file is the same as the name of the source package,
except the extension is \code{.zip}.
}
\keyword{internal}
