% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{remotes_add_remotes}
\alias{remotes_add_remotes}
\title{Add remotes to manage}
\usage{
remotes_add_remotes(self, private, specs)
}
\arguments{
\item{self}{Self.}

\item{private}{Private self.}

\item{specs}{Character vector, remotes specifications.}
}
\description{
We also parse them to make sure that they are valid,
but don't actually try to resolve them just yet.
}
\keyword{internal}
