% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo$add}
\alias{repo$add}
\title{Add a package to a repository}
\usage{
repo$add(file, ..., path = ".")
}
\arguments{
\item{file}{Package file.}

\item{...}{Ignored currently.}

\item{path}{Path to repository. Must contain a \code{PACKAGES} file.}
}
\description{
It does not check if any version of the package is already
in the repository. If you want to \emph{update} a package, use
\code{repo_update()}.
}
\keyword{internal}
