% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo$delete}
\alias{repo$delete}
\title{Delete packages from repository metadata}
\usage{
repo$delete(package, ..., path = ".")
}
\arguments{
\item{package}{Package name.}

\item{...}{Other fields to match, they must be named. Matching is
case insensitive.}

\item{path}{Path to repository. Must contain a \code{PACKAGES} file.}
}
\description{
All matching packages will be removed.
}
\details{
It does not delete the files themselves.
}
\keyword{internal}
