% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo$list}
\alias{repo$list}
\title{List packages in a reporitory}
\usage{
repo_list(..., path = ".")
}
\arguments{
\item{...}{Ignored currently.}

\item{path}{Path to repository. Must contain a \code{PACKAGES} file.}
}
\value{
Data frame of package data.
}
\description{
List packages in a reporitory
}
\keyword{internal}
