% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo$update}
\alias{repo$update}
\title{Update a package in a repository}
\usage{
repo$update(file, ..., path = ".")
}
\arguments{
\item{file}{Package file.}

\item{...}{Ignored currently.}

\item{path}{Path to repository. Must contain a \code{PACKAGES} file.}
}
\description{
Previous version of the same package are removed. In particular,
it removes packages with matching:
\itemize{
\item package name (\code{Package} field),
\item R version (\code{Rversion} field),
\item same OS (\code{OS} field), or no \code{OS} field,
\item same architecture (\code{Arch} field), or not \code{Arch} field.
}
}
\keyword{internal}
