% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo$update_gh}
\alias{repo$update_gh}
\title{Update a file in a package metadata, stored on GitHub}
\usage{
repo$update_gh(repo, subdir, file)
}
\arguments{
\item{repo}{GitHub slug, e.g. \code{r-hub/repos}.}

\item{subdir}{Subdirectory in the GitHub repository, where the R package
metadata should be updated. It must exist in the repository.
If it does not have \verb{PACKAGES*} files, then they will be created.}

\item{file}{Package file to add. The file will \emph{not} be added to the
repository.}
}
\description{
\enumerate{
\item Clones the GitHub repository.
\item Calls \code{repo_update()} with \code{file}, in the \code{subdir} directory.
\item Adds and commits changes.
\item Pushes the git repository to GitHub.
If the push fails, then it resets the git repository, pulls it from
GitHub and tries the update process again, until the push succeeds.
}
}
\details{
This function needs command line git installed.

It sets up a \code{cache} credential helper, so the \verb{git push} works
without interaction with the user.
}
\keyword{internal}
