% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{sysreqs_db_list}
\alias{sysreqs_db_list}
\title{List contents of the system requirements DB, for a platform}
\usage{
sysreqs_db_list(platform = NULL)
}
\arguments{
\item{platform}{System requirements platform.}
}
\value{
Data frame with columns:
\itemize{
\item \code{name}: cross platform system dependency name in the database.
\item \code{patterns}: one or more regular expressions to match to
\code{SystemRequirements} fields.
\item \code{packages}: one or more system package names to install.
\item \code{pre_install}: command(s) to run before installing the packages.
\item \code{post_install}:: command(s) to run after installing the packages.
}
}
\description{
List contents of the system requirements DB, for a platform
}
\seealso{
Other system requirements functions: 
\code{\link{sysreqs_check_installed}()},
\code{\link{sysreqs_db_match}()},
\code{\link{sysreqs_db_update}()},
\code{\link{sysreqs_install_plan}()},
\code{\link{sysreqs_is_supported}()},
\code{\link{sysreqs_list_system_packages}()},
\code{\link{sysreqs_platforms}()}
}
\concept{system requirements functions}
