% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{sysreqs_is_supported}
\alias{sysreqs_is_supported}
\title{Check if a platform has system requirements support}
\usage{
sysreqs_is_supported(platform = NULL)
}
\arguments{
\item{platform}{System requirements platform.}
}
\value{
Logical scalar.
}
\description{
Check if a platform has system requirements support
}
\examples{
sysreqs_is_supported()
}
\seealso{
The \code{sysreqs_platform} \link[=pkgdepends-config]{configuration option}.

Other system requirements functions: 
\code{\link{sysreqs_check_installed}()},
\code{\link{sysreqs_db_list}()},
\code{\link{sysreqs_db_match}()},
\code{\link{sysreqs_db_update}()},
\code{\link{sysreqs_install_plan}()},
\code{\link{sysreqs_list_system_packages}()},
\code{\link{sysreqs_platforms}()}
}
\concept{system requirements functions}
