% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-dirs.R
\name{user_metadata_cache_dir}
\alias{user_metadata_cache_dir}
\title{Cache directory for the CRAN (or other) metadata}
\usage{
user_metadata_cache_dir()
}
\value{
path
}
\description{
We want \code{pkgdepends} to work with concurrent R sessions, so we need
to keep the metadata consistent. So we use a lock, and make a copy
of the directory, to avoid holding the lock for a long time.
This way we don't hold up other processes. Then we update the cache
after acquiring a write lock.
}
\details{
For \code{repoman} we do not use these caches, but instead cache everythin
within the repository.
}
\keyword{internal}
