% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-api.R
\name{cache_summary}
\alias{cache_summary}
\alias{cache_list}
\alias{cache_find}
\alias{cache_get_file}
\alias{cache_delete_files}
\alias{cache_add_file}
\title{Functions to query and manipulate the package cache}
\usage{
cache_summary(cachepath = NULL)

cache_list(cachepath = NULL)

cache_find(cachepath = NULL, ...)

cache_get_file(cachepath = NULL, target, ...)

cache_delete_files(cachepath = NULL, ...)

cache_add_file(cachepath = NULL, file, relpath = dirname(file), ...)
}
\arguments{
\item{cachepath}{Path of the cache. By default the cache directory is in
\code{R-pkg}, within the user's cache directory.
See \code{\link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir()}}.}

\item{...}{Extra named arguments to select the package file.}

\item{target}{Path where the selected file is copied.}

\item{file}{File to add.}

\item{relpath}{The relative path of the file within the cache.}
}
\description{
\code{cache_summary()} returns a short summary of the state of the cache,
e.g. the number of files and their total size. It returns a named list.
}
