% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-cache.R
\name{get_package_from}
\alias{get_package_from}
\title{Get package from cache, or download it asynchronously}
\usage{
get_package_from(cache, urls, cache_dir, target)
}
\arguments{
\item{cache}{\code{package_cache} instance.}

\item{urls}{Character vector, list of candidate urls.}

\item{target_file}{Where to save the file.}
}
\value{
Download status.
}
\description{
Currently, even if we take it from the cache, we check the etag
of the url. TODO: This should change in the future, and we should trust
at least CRAN source packages.
}
\keyword{internal}
