% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-cache.R
\name{get_package_from}
\alias{get_package_from}
\title{Get package from cache, or download it asynchronously}
\usage{
get_package_from(cache, urls, target_dir, target, progress_bar = NULL,
  direct = NULL, metadata = list())
}
\arguments{
\item{cache}{\code{package_cache} instance.}

\item{urls}{Character vector, list of candidate urls.}

\item{target_dir}{Directory to place the file in.}

\item{target}{Path to target file, within the target directory.}

\item{progress_bar}{The progress bar object to update.}

\item{direct}{Whether the package/ref was directly specified
(or a dependency).}

\item{metadata}{Extra data to add to the installed package.}
}
\value{
Download status.
}
\description{
Get package from cache, or download it asynchronously
}
\keyword{internal}
