% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-protocol.R
\name{git-protocol}
\alias{git-protocol}
\title{git protocol notes, for developers}
\description{
Assumptions, they might be relaxed or checked for later:
\itemize{
\item The server must speak the smart protocol, version 2.
\item We use HTTP transport, not SSH.
\item The server should have the \code{shallow} capability.
\item The server should have the \code{filter} capability.
}
}
\details{
Improvements needed:
\itemize{
\item Tests.
\item Use async HTTP.
\item Optionally send authorization.
}
}
\keyword{internal}
