% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-remotes.R
\name{github_username_rx}
\alias{github_username_rx}
\title{Match a GitHub username}
\usage{
github_username_rx()
}
\description{
\itemize{
\item may only contain alphanumeric characters or hyphens
\item cannot have multiple consecutive hyphens
\item cannot begin or end with a hyphen
\item maximum 39 characters
}

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\details{
Based on \url{https://github.com/shinnn/github-username-regex}
}
\keyword{internal}
