% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-binary.R
\name{install_binary}
\alias{install_binary}
\title{Install an R binary package}
\usage{
install_binary(
  filename,
  lib = .libPaths()[[1L]],
  metadata = NULL,
  quiet = NULL
)
}
\arguments{
\item{filename}{filename of built binary package to install}

\item{lib}{library to install packages into}

\item{metadata}{Named character vector of metadata entries to be added
to the \code{DESCRIPTION} after installation.}

\item{quiet}{Whether to suppress console output.}
}
\description{
Uncompress an R binary package into a package library.

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css"><script src="../doc/assets/rd.js"></script>}}
}
