% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-plan.R
\name{install_package_plan}
\alias{install_package_plan}
\title{Perform a package installation plan}
\usage{
install_package_plan(
  plan,
  lib = .libPaths()[[1]],
  num_workers = 1,
  cache = NULL
)
}
\arguments{
\item{plan}{Package plan object, a data frame, see
\link[=install_plans]{'Installation plans'} for the format.}

\item{lib}{Library directory to install to.}

\item{num_workers}{Number of worker processes to use.}

\item{cache}{Package cache to use, or \code{NULL}.}
}
\value{
Information about the installation process.
}
\description{
See \link[=install_plans]{'Installation plans'} for the details and the format.
}
