% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-installed.R
\name{lib_status}
\alias{lib_status}
\title{Status of packages in a library}
\usage{
lib_status(library = .libPaths()[1], packages = NULL)
}
\arguments{
\item{library}{Path to library.}

\item{packages}{If not \code{NULL}, then only these packages are shown.}
}
\value{
Data frame (tibble) the contains data about the packages
installed in the library.
}
\description{
Query data of all packages in a package library.
}
