% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-remotes.R
\name{parse_remotes}
\alias{parse_remotes}
\title{Parse package location specifications}
\usage{
parse_remotes(specs, remote_types = NULL, ...)
}
\arguments{
\item{specs}{character vector}
}
\value{
List of parsed specification.
}
\description{
Parse package location specifications
}
