% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\name{pkg_rx}
\alias{pkg_rx}
\title{A set of handy regular expressions related to R packages}
\usage{
pkg_rx()
}
\value{
A named list of strings.
}
\description{
If you use these in R, make sure you specify \code{perl = TRUE}, see
\code{\link[base:grep]{base::grep()}}.
}
\details{
Currently included:
\itemize{
\item \code{pkg_name}: A valid package name.
\item \code{type_cran}: A \verb{cran::} package reference.
\item \code{type_bioc}: A \verb{bioc::} package reference.
\item \code{type_standard}: A \verb{standard::} package reference.
\item \code{type_github}: A \verb{github::} package reference.
\item \code{type_local}: A \verb{local::} package reference.
\item \code{type_deps}: A \verb{deps::} package reference.
\item \code{type_installed}: An \verb{installed::} package reference.
\item \code{github_username}: A GitHub username.
\item \code{github_repo}: A GitHub repository name.
\item \code{github_url}: A GitHub URL.
}
}
\examples{
pkg_rx()
}
