% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-progress-bar.R
\name{pkgplan__update_progress_bar}
\alias{pkgplan__update_progress_bar}
\title{Update the progress bar data}
\usage{
pkgplan__update_progress_bar(bar, idx, data)
}
\arguments{
\item{bar}{The progress bar.}

\item{idx}{The row index in the download data table.}

\item{data}{Data about this row, either \code{"done"}, meaning that the
download has completed, or a list with entries \code{current} and \code{total},
measured in the number of bytes. If \code{total} is unknown, then it is
set to zero.}
}
\description{
This is triggered by libcurl.

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\keyword{internal}
